+++ frontend/apps/marketing/src/middleware/utils/getCachedRedirectResponse.ts
@@ -1,5 +1,7 @@
 import {NextResponse} from 'next/server';
 
+import {STALE_WHILE_REVALIDATE_ONE_HOUR} from '@/cache/constants';
+
 /**
  * Sets the Cache-Control header on a redirect response to cache the redirect on the CDN.
  * @param url The URL to redirect to.
@@ -8,10 +10,7 @@ export function getCachedRedirectResponse(url: string | URL) {
   const response = NextResponse.redirect(url);
 
   // Cache the redirect on Cloudfront
-  response.headers.set(
-    'Cache-Control',
-    's-maxage=3600, stale-while-revalidate=31535100',
-  );
+  response.headers.set('Cache-Control', STALE_WHILE_REVALIDATE_ONE_HOUR);
 
   return response;
 }
